Add reading of response key on callback timeout. lwa-issues #52, #55
Increase default user timeout on raw command.
Response key now being sent to service to use for response thus decoupling the
two. On occasion, a race between callback registration and sending the response
to the key occurs. Logic added to check the response key if the callback times
out. Also added a new Exception when the service does not respond.

Signed-off-by: rh <githubrh@pm.me>